Configure RBAC for cluster admin access
Grant members of the `cluster-admins` group both `cluster-reader` and
`sudoer` access. This permits members of this group to see most
cluster resources, and to impersonate other users, including the
`system:admin` user for full cluster-admin access.

From the command line, you can use the `--as` option to impersonate
the admin user:

   oc --as system:admin create ns example