Add udev rules to identify mellanox network interfaces
We disabled predictable network device names in the previous commit,
which means we're now at the mercy of the order in which the kernel
scans network devices. If a card is added or removed from the system,
what we know as "eth2" may no longer be the same interface it is now.

To solve this, we add udev rules that match the network interfaces
based on characteristics such as their PCI bus/device/function.