Configure OpenShift image registry
Now that we have storage we can enable the image registry. Because
we're limited to Ceph RBD storage, we only have RWO [1] volumes, which
means we can only have a single replica for the registry service.

It is possible to use an S3 backend for the registry (which would
allow us to have multiple replicas), but it looks as if getting things
set up to use a bucket from ODF [2] will be tricky.

[1]: https://kubernetes.io/docs/concepts/storage/persistent-volumes/#access-modes
[2]: https://www.redhat.com/en/technologies/cloud-computing/openshift-data-foundation