Fix custom logo
The instructions for creating a custom logo [1] ask us to create a
ConfigMap in the openshift-config namespace. It looks as the console
operator then *copies* that into the openshift-console namespace, renaming
it to `custom-logo`. This copy preserves all the labels, including those
added by ArgoCD, which causes ArgoCD to identify the `custom-logo` object
as something that should be deleted.

This commit ignores the documentation and instead directly populates
the `custom-logo` ConfigMap in the openshift-console namespace.

[1]: https://docs.openshift.com/container-platform/4.10/web_console/customizing-the-web-console.html