Increase resource limits for patch operator
While experimenting with the patch operator, the controller was repeatedly
getting OOMKilled due to hitting its declared resource limits. Based on [1]
and [2], it looks like we can change the resource limits of the controller
by modifying the Subscription.

This commit attempts to increase the limits to 1G memory and 1.5 cpu
shares.

[1]: https://github.com/redhat-cop/patch-operator/issues/48
[2]: https://github.com/operator-framework/operator-lifecycle-manager/blob/master/doc/design/subscription-config.md#resources