Enable SkipDryRunOnMissingResource for some resources
When applying manifests, ArgoCD performs a dry-run to ensure that all the
resources types referred to in the manifests will apply correctly. When we
are installing resources that require CRDs provided by an operator that is
being installed in the same operation, the dry-run will fail because those
resources do not exist in either the cluster or in the applied manifests.

This commits instructs ArgoCD to skip the dry run for some resource
types, per documentation in [1].

[1]: https://argo-cd.readthedocs.io/en/stable/user-guide/sync-options/#skip-dry-run-for-new-custom-resources-types