Increasing the Loki ingestionRate and ingestionRateBurstSize
Increasing the Loki ingestionRate to 10MB and ingestionRateBurstSize of
10Mb. The ingestion rate was not large enough after the logging stopped
working while switching to the public Loki Gateway URL before I configured
the right TLS certificates and key in vault.  After I added a tls.crt,
tls.key and the right ca-bundle.crt for the new route in the
lokistack-gateway-bearer-token secret in vault the OpenShift Logging Pods
were able to connect. I restarted the openshift-logging pods and it started
working except it had too many logs to ingest at one time. So please review
a PR to increase the ingestion limit that get the logs going again.