Moving the lokistack-gateway-bearer-token to the logging app
While deploying the `logging` ArgoCD app to prod, an error occured: secret "lokistack-gateway-bearer-token" not found. This is because the Secret is currently defined in the `loki` app, which is only deployed on the infra cluster. This commit will move the lokistack-gateway-bearer-token Secret to the `logging` app instead where it will be deployed in both the infra and prod cluster to fix cluster logging in prod.