Configure bond0 interface
We will ultimately have redundant connections from out nodes to the switch.
This commits configures bond0 on across interfaces eno1 and eno2. All
active nodes currently share the same interface names; we will need to
revisit this configuration if we introduce nodes on which the interfaces
differ; see the nerc-ocp-infra cluster for one approach.