Ask argocd to ignore custom-logo resource
When providing a custom logo to the console application, the console
operator copies the configmap from openshift-config to the
openshift-console namespace, preserving all labels. This makes argocd think
the resource needs to be tracked.

With this commit we attempt to instruct argocd to ignore the extraneous
resource by adding the appropriate argocd.argoproj.io/compare-options
annotation [1].

[1]: https://argo-cd.readthedocs.io/en/stable/user-guide/compare-options/