This was the causing the groups to not sync in either cluster.
The operator on both clusters complained about not finding the secret,
though I am not sure why the groups exist in the infra cluster.

I will also update the token in vault to match it.

It would be easier to update the credentialSecret in GroupSync definition,
but I like the name `github-group-sync-token` slightly more than `github-ocp-on-nerc`.