Remove custom routes for public ingress
The routes added by these files are technically correct, but due to the way
loadbalancer addresses are handled the rules don't apply to traffic in a
useful fashion.

Traffic coming from a service is modified by a masquerade rule in the
POSTROUTING chain to have the correct source address. Because this happens
*after* a routing decision has already been made, the outbound request
never matches the routing policy rules defined by these files.