Patch external ingress service via argocd post-sync hook
When we create an IngressController, the IngressController creates a
Service. We need to modify that service resource to add both an annotation
and a loadBalancerIP, but ArgoCD doesn't allow us to patch existing
resources.

The best we can do is apply patches using an ArgoCD post-sync hook [1].

This commit adds a post-sync Job that takes care of patching the Service
resource. This job will run after every sync operation.

[1]: https://argo-cd.readthedocs.io/en/stable/user-guide/resource_hooks/