Upgrading ACM and Loki Operators
We ran into an issue with Observability where the prod cluster was no longer reporting metrics.
Observability is back online thanks to a suggestion from @jakobo from the CoreOS forum-acm-cluster-lifecycle Slack channel.
After deleting the multicluster-engine Subscription and CSV, deleting the openshift-monitoring pods, and deleting the ACM Subscription, CSV, and MultiClusterObservability CRD again, the infra and prod cluster metrics are back in Observability.

We are also migrating from the candidate channel of the LokiOperator to the stable-5.5 channel.