Add labels to nodes required by OCS/ODF for setting up noobaa
This uses the patch operator to add labels to the nodes of nerc-ocp-infra
and nerc-ocp-prod clusters.

A clusterrole to allow adding labels to node is added to the common overlay.

A serviceaccount is created in the openshift-storage namespace and bound by
a clusterrolebinding to the node-labeler clusterrole.

The clusterrolebinding and the service account is included via the odf bundle.

The actual patches are in the overlays.