Move ArgoCD "SkipDryRunOnMissingResource" patches to a component
Back in bb8a93dc40 we attempted to move the "SkipDryRunOnMissingResource"
patches for ArgoCD into a separate file, but that didn't work because
(a) patches only effect resources generated by the `kustomization.yaml` in
which they are included, and (b) we had moved the patches into an isolated
subdirectory, so they were effectively a no-op.

But there is a better way! By making them a component [1] rather than a
resource, they are effectively included in the current kustomization.yaml.
This lets us both reduce the clutter in the nerc-ocp-prod overlay *and*
allows us to share the same set of patches in other overlays.

[1]: https://github.com/kubernetes/enhancements/blob/master/keps/sig-cli/1802-kustomize-components/README.md