Install Apache Solr Operator Helm Chart Repository and CRDs to prod
The Smart Village Collaboratory project would like to install the Apache Solr Operator and CRDs to NERC. There are two important parts here. The Apache Zookeeper cluster for running Solr as a clustered cloud, as well as our smart village application which is a Vert.x application that is also clustered between pods with the Apache Zookeeper cluster. The second part is Apache Solr which is the search engine for the Smart Village application. The search engine powers the API for extremely fast querying, pagination, filtering, faceting, pivoting, sorting and more. This allows for building custom analytics and dashboards that are extremely powerful. You can find the CRDs for Apache Solr here, and the Helm Chart Repository for Apache Solr here.

closes https://github.com/OCP-on-NERC/operations/issues/110