Adding a Metric to Observability, so that the container information will be monitored.
My intern project focuses on scanning the vulnerabilities in the containers running in the cluster. One step is to retrieve the images used by the containers running in the cluster and then we will analyse the dependencies in the image. Adding `kube_pod_container_info` to Observation exposes the container image sources for us to further introspection.