Add udev rules for gpu nodes
We configure the bond0 interface to span interfaces named "nic1" and
"nic2", and we rely on udev rules to assign these names to the appropriate
interfaces. There were no such rules for the GPU nodes (wrk-{10,11}) on the
test cluster.

This commit adds the necessary udev rules.

Part of: ocp-on-nerc/operations#170