add odh dashboard config for prod cluster
This config disables the notebook controller which disables the classic
"tiles". The classic tiles use a shared/global namespace which is not
desirable on the prod cluster. On the prod cluster, each user is
confined to allocations/namespaces that were approved and provisioned
via coldfront.

Also disables "enablement", "support", and "info" features related to
classic "tiles".