Adding OpenShift Custom Metrics Autoscaler Operator to OpenShift Prod
The Ecological Forecasting project wishes to scale up their consumers of
rabbitmq queues based on the number of messages in the queue. This is an
out-of-the-box feature of the OpenShift Custom Metrics Autoscaler Operator. The
ScaledObject CRD provided by this operator will allow us to specify the
Deployment we wish to scale, the min and max replica count, and the RabbitMQ
queue connection details.