prod: fix NIC device names via udev rules
Previously we had homogeneous worker node hardware, however, with the
addition of 2x new NVIDIA V100 worker hosts this is no longer the case.
These new hosts have different NIC devices than the rest of the worker
nodes.

This adds the following new machineconfigs in order to map all of the
NICs in the cluster to nic1 and nic2 devices:

- disable-net-ifnames
- custom-udev-rules

The configure-bond0 machineconfig has been updated to use nic1 and nic2
vs the old eno1 and eno2 device names. Previously, this machineconfig
was only being applied on the worker nodes. This patch also extends that
configuration to the controllers in order to setup an LACP bond on those
hosts.