Patch vault path for route53 key in each overlay
While we have a "nerc-certificate-issuer" component, it was only used on
nerc-ocp-prod. In order to use this on other clusters, we'll need to patch
the ExternalSecret resource with a new key path.

This commit implements that behavior for nerc-ocp-prod to keep things
consistent as we add new clusters.

Part-of: nerc-project/operations#285