Use ClusterIssuers instead of namedspaced Issusers on obs cluster (#361)
* Fix typo in staging issuer configuration

The prod and staging issuers should probably be using different secrets for
storing registration information.

* Add dns-01 based ClusterIssuer

This is useful on clusters on which we are not running arbitrary user
workloads and thus are less concerned about granting the ability to request
`*.nerc.mghpcc.org` certificates in any namespace.

* Replace namespaced Issuers with ClusterIssuers on obs cluster

We're not running user workloads on this cluster so we can simplify things
by using a single ClusterIssuer rather than multiple namespaced Issuers.