obs cluster loki-operator upgrade from 5.6 to 5.7
The Loki Operator Subscription on the obs cluster is unable to install
because of an error: constraints not satisfiable: no operators found in
channel stable-5.6 of package loki-operator in the catalog referenced by
subscription loki-operator, subscription loki-operator exists.

We will upgrade Loki to version 5.7 on the obs cluster because
version 5.6 is no longer supported.