Move scripts into ConfigMaps
Storing the backup scripts in ConfigMaps lets us store them as
individual files in the repository, which makes it much easier to
apply formatting, syntax checking, and other linting behavior when
editing the files locally.

We use kustomize to assemble the individual files into a ConfigMap,
and then mount that into the Task.

This commit includes fixes for all issues reported by shellcheck [1].

[1]: https://github.com/koalaman/shellcheck