Add vault operator
The vault operator [1] adds a controlled and CRDs for creating vault
clusters. This commit adds only the operator manifests; it does not
create any vault instances.

The operator takes care of creating new vault clusters, initializing
them, storing the unseal keys in a kubernetes Secret, and
automatically unsealing the vault when necessary.

[1]: https://bank-vaults.dev/docs/operator/

Part-of: nerc-project/operations#446
Part-of: nerc-project/operations#439