Make a generic, reusable "ops" role
This introduces a generic "ops" rbac that can be used to create one or
more sudo-capable read-only cluster roles. For example, we can use
this framework to create our existing "nerc-ops" roles, and we can
also use it to create a per-cluster "local-ops" role (permitting a
group to have elevated privileges on one cluster but not on others).