Add additional vault redirect uris to dex configuration
The new vault deployment can be reached at the shorter hostname
`vault.apps.nerc-ocp-infra.rc.fas.harvard.edu`, but that URL wasn't
registered with Dex so it was not possible to log in with that URL when
using OIDC (token login worked fine).

This commit adds the necessary redirect URIs to the dex configuration.