Configure certificate trust for infra services accessing the vault
Now that most vault access is over tls, we need to configure certificate
trust between services on the infra cluster and the vault (services
external to the infra cluster see the default ingress certificate rather
than the internal vault certificate).

This commit modifies the external secrets operator and the backup job so
that they can interact with vault's https endpoint.