Manage vault configuration
This commit adds our vault configuration to the repository.

We use yq [1] to merge multiple YAML files in the final configuration
expected by the bank-vaults cli [2], which we then use to apply the
configuration to the vault.

The configuration is applied by an ArgoCD PostSync hook [3] whenever
there are updates to the repository.

[1]: https://github.com/mikefarah/yq
[2]: https://github.com/bank-vaults/bank-vaults
[3]: https://argo-cd.readthedocs.io/en/stable/user-guide/resource_hooks/

Closes: nerc-project/operations#449