Add default Keycloak openid scope to obs NERC realm and clients
This relates to the latest updates in the prom-keycloak-proxy that check
the userinfo realm endpoint for the username of the user making requests
with their Authorization header bearer token. We add the openid scope
to all clients in the realm, which is required for using the userinfo
endpoint to check the username of an access token.