Add schedule to groupsync resources
This adds a schedule to our GitHub group sync configuration; from the
documentation [1]:

> If a schedule is not provided, synchronization will occur only when the
> object is reconciled by the platform.

This means that previously, the group sync operator was not actually
synchronization groups other than (a) when first starting up or (b) when
changes were made to the group sync configuration.

Since we're using this configuration pretty much everywhere, this commit
also relocated the group sync configuration into cluster-scope/base and
introduces a new group-sync-github bundle that includes it by default.

[1]: https://github.com/redhat-cop/group-sync-operator#scheduled-execution

Closes: nerc-project/operations#653