ocp-test: add missing kubevirt manifests
The default install of the kubevirt/cnv operator seems to be missing
several manifests that should be in the default namespace. Adding them
to the nerc-ocp-test overlay for now until we hear back from RH support
on whether these should come out of the box or if we're expected to
configure these ourselves.