Grant project-robbie-b4784c access to anyuid scc
This grants the `robbie-job-runner` in project `project-robbie-b4784c` use
of the `anyuid` scc. This allows Matt Brewster's container(s) to run using
predefined UIDs.

This PR also modifies the previous change for this project such that we are
using RoleBindings instead of ClusterRoleBindings to grant permissions.
This ensures that the RBAC is clearly associated with the namespace
(although it should not introduce any operational changes).

Closes: nerc-project/operations#734