nerc-ocp-test: add unique acceleratorProfiles for A100 vs V100 GPUs
Addresses: https://github.com/nerc-project/operations/issues/767

Prior to this commit, there was no way to distinguish between different GPU node types in RHOAI
The current acceleratorProfile simply looks for any node with the label nvidia.com/gpu, whether that schedules the workload in A100 or V100 is up to chance

The tolerations in the acceleratorProfiles in this commit inject tolerations into RHOAI workloads based on the the choice of NVIDIA GPU from the RHOAI dashboard.

NOTE: AcceleratorProfiles do not taint the nodes, they only locate gpus and inject tolerations to workloads.