Create ocp-test cluster health dasboard in ocp-obs cluster
This commit adds the yaml for a cluster health dashboard for the ocp-test cluster. This can be a resource which help cluster admins in
NERC clusters to more quickly and efficiently recognize  and diagnose issues and provide a bird's eye view of the cluster.
We can add this dashboard to all clusters, if folks find it useful.