Add argocd post-sync hook to patch dns operator
This adds a job that discovers the address of the custom dns service and
patches the dns operator appropriately. This will run automatically
whenever argocd pushes changes to the cluster.

The job will also run when you `kubectl apply` these manifests, but in
this case you will need to manually remove the job first if it has
previously been created (and if you are pushing a modified version).