console: add managementState to base manifest
On some clusters the console operator gets degraded with the following
error:

ManagementStateDegraded: Unsupported management state "" for console
operator...

This is due to missing the managementState: managed field in the console
spec. See https://access.redhat.com/solutions/7002156

Oddly enough, we're missing this parameter in the manifests of most
clusters except for nerc-ocp-test and they don't appear to have this
issue. Checking on the manifests in various clusters that don't have
this in their console spec in this repo, they still seem to have
managementState: managed.

Adding this to force the situation globally.