vault: use full registry domain in image ref
After upgrading the infra cluster to 4.19, the vault pods were failing
to pull the image due to using the wrong registry (RedHat) and for some
reason weren't failing over to the final default registry of docker.io
in the unqualified-search-registries list.

This explicitly pulls from docker hub instead of relying on the
unqualified search registries list when using short names. This is also
recommended for security reasons in the RH OCP docs.