Upgrades to AI Telemetry to zookeeper-operator and solr-operator (#824)
* Add Kafka support to AI Telemetry

This is used for event driven messaging from other microservices and
tasks that take a long time, like bare metal fulfillment.

* Replace bitnami zookeeper with zookeeper-operator

Because bitnami images are now proprietary, we are migrating to the open
source zookeeper-operator in OpenShift instead.

* Migrate from bitnami solr to solr-operator

Because bitnami images are now proprietary, we are migrating to the open
source solr-operator on OpenShift instead. Because the solr-operator
upstream does not support OpenShift yet, it is pointed to my fork where
I have contributed OpenShift support to solr-operator pull request [1].

[1] https://github.com/apache/solr-operator/pull/706

* Adding setup jobs for ai-telemetry for solr migration

Because we are migrating to the solr-operator, there is a solr setup job
and db-to-solr-sync job that are run to migrate the persistent data in
the database to be cached in the Solr search engine which is used by the
AI Telemetry JSON REST API.

* Updating the prom-keycloak-proxy to 2.3.0

This resolves and import error for handling null metrics results.  Also
fixes the PROXY_PROMETHEUS_CA_CERT and PROXY_PROMETHEUS_TLS_CERT
environment variables.