ocp-prod: add clusterrolebinding for listing crds (#834)
It turns out customresourcedefinitions are cluster-scoped which means we
can't give access to users via a namespace rolebinding.

This patch adds a new bundle to include both a clusterrole and
clusterrolebinding for allowing all authenticated users read-only access
to customresourcedefinitions.

This patch also removes the aggregate-to-edit label from the clusterrole
given that we have to bind this at the cluster-scope (ie not from a
namespace rolebinding to the edit clusterrole).

Finally, it adds the bundle to the ocp-prod overlay.