Upgrade to latest prom-keycloak-proxy 2.4.0 features. Add init container for loading AI Telemetry courses. Add persistent-metrics-zookeeper AI Telemetry custom metrics. Adding a missing solr serviceaccount. Adding kube_namespace_status_phase and kube_namespace_labels.  (#836)
* Adding kube_namespace_status_phase and kube_namespace_labels

These metrics are useful for reporting on namespaces stuck deleting, and
for namespace labels. These metrics account for all namespaces in the
cluster which we can't already do.

* Adding a missing solr serviceaccount

* Upgrade to latest prom-keycloak-proxy 2.4.0 features

* Add init container for loading AI Telemetry courses

This init container pulls the AI Telemetry courses and automatically
imports them when the server starts up.

* Add persistent-metrics-zookeeper AI Telemetry custom metrics

This is a simple python application connected to zookeeper, which is a
distributed counter for creating accurate, persistent custom metrics.